[metasrv] refactor: move network message conversions into metasrv/meta_service/network.rs
