+++ metasrv/src/raft/state_machine/applied_state.rs
@@ -22,9 +22,6 @@ use common_store_api_sdk::storage_api_impl::DataPartInfo;
 use serde::Deserialize;
 use serde::Serialize;
 
-use crate::meta_service::RaftMes;
-use crate::meta_service::RetryableError;
-
 /// The state of an applied raft log.
 /// Normally it includes two fields: the state before applying and the state after applying the log.
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
@@ -142,39 +139,6 @@ impl From<(Option<SeqValue<KVValue>>, Option<SeqValue<KVValue>>)> for AppliedSta
         AppliedState::KV {
             prev: v.0,
             result: v.1,
-        }
-    }
-}
-
-// === from and to transport message
-
-impl From<AppliedState> for RaftMes {
-    fn from(msg: AppliedState) -> Self {
-        let data = serde_json::to_string(&msg).expect("fail to serialize");
-        RaftMes {
-            data,
-            error: "".to_string(),
-        }
-    }
-}
-impl From<Result<AppliedState, RetryableError>> for RaftMes {
-    fn from(rst: Result<AppliedState, RetryableError>) -> Self {
-        match rst {
-            Ok(resp) => resp.into(),
-            Err(err) => err.into(),
-        }
-    }
-}
-
-impl From<RaftMes> for Result<AppliedState, RetryableError> {
-    fn from(msg: RaftMes) -> Self {
-        if !msg.data.is_empty() {
-            let resp: AppliedState = serde_json::from_str(&msg.data).expect("fail to deserialize");
-            Ok(resp)
-        } else {
-            let err: RetryableError =
-                serde_json::from_str(&msg.error).expect("fail to deserialize");
-            Err(err)
         }
     }
 }
